Feature/shiperus/merge ij branch to sellerapp

+++ features/travel/hotel/src/main/java/com/tokopedia/hotel/roomlist/data/model/HotelAddCartParam.kt
@@ -22,7 +22,7 @@ data class HotelAddCartParam(
 
         @SerializedName("propertyID")
         @Expose
-        val propertyId: Int = 0,
+        val propertyId: Long = 0,
 
         @SerializedName("rooms")
         @Expose
